[mage-1761] 1) Fix infinite retry loop when user locations fail to push to server 2) Refactor location classes to differentiate between the active user's location tracking vs the event's user locations
